cdd ConfirmDel & ete eÃ & Add Quantity Low Stock Alert

+++ app/src/main/java/com/alimrhjmohammadshahoud/pharmacymanagementapp/MedicineListActivity.java
@@ -5,6 +5,7 @@ import android.app.AlertDialog;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.EditText;
+import android.widget.TextView;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
@@ -37,7 +38,7 @@ public class MedicineListActivity extends AppCompatActivity {
         fabAddMedicine = findViewById(R.id.fab_add_medicine);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
-        allMedicines.add(new Medicine(1, "Panadol", 1, 5, 6.0));
+        allMedicines.add(new Medicine(1, "Panadol", 1, 3, 6.0));
         allMedicines.add(new Medicine(3, "Aspirin", 3, 5, 6.0));
         allMedicines.add(new Medicine(4, "Vitamin C", 1, 10, 12.0));
 
@@ -77,7 +78,7 @@ public class MedicineListActivity extends AppCompatActivity {
 
                         Medicine newMed = new Medicine(id, name, companyId, qty, price);
                         adapter.addMedicine(newMed);
-                        adapter.notifyItemInserted(filteredMedicines.size() - 1);                        adapter.notifyItemInserted(filteredMedicines.size() - 1);
+                        adapter.notifyItemInserted(filteredMedicines.size() - 1);
                     } catch (Exception e) {
                     }
                 })
@@ -95,6 +96,7 @@ public class MedicineListActivity extends AppCompatActivity {
                     try {
                         double newPrice = Double.parseDouble(editMedicinePrice.getText().toString());
                         adapter.changePriceMedicine(med,newPrice);
+                        adapter.notifyItemChanged(position);
                     } catch (Exception e) {
                         e.printStackTrace();
                     }
@@ -102,10 +104,40 @@ public class MedicineListActivity extends AppCompatActivity {
                 .setNegativeButton("Cancel", null)
                 .show();
     }
-
-    public void deleteMedicine(Medicine med) {
-        int position = filteredMedicines.indexOf(med);
-        adapter.deleteMedicine(med,position);
+    @SuppressLint("MissingInflatedId")
+    public void showAddQuantityToMedicine(Medicine med, int position) {
+        View view = getLayoutInflater().inflate(R.layout.activity_dialog_add_quantity_to_medicine, null);
+        EditText editMedicineQuantity = view.findViewById(R.id.add_Quantity);
+        editMedicineQuantity.setText(String.valueOf(0));
+        new AlertDialog.Builder(this)
+                .setTitle("Enter the quantity to add")
+                .setView(view)
+                .setPositiveButton("Update", (dialog, which) -> {
+                    try {
+                        int quantity_Added = Integer.parseInt(editMedicineQuantity.getText().toString());
+                        int new_Quantity = quantity_Added + med.getQuantity();
+                        adapter.addToQuantity(med,new_Quantity);
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
+                })
+                .setNegativeButton("Cancel", null)
+                .show();
+    }
+    public void showConfirmeToDeleteMedicine(Medicine med, int position) {
+        View view = getLayoutInflater().inflate(R.layout.activity_confim_delete_medicine, null);
+        new AlertDialog.Builder(this)
+                .setTitle("Are you sure you want to delete this Medicine?")
+                .setView(view)
+                .setPositiveButton("Confirme", (dialog, which) -> {
+                    try {
+                        adapter.deleteMedicine(med,position);
+                    } catch (Exception e) {
+                        e.printStackTrace();
+                    }
+                })
+                .setNegativeButton("Cancel", null)
+                .show();
     }
 }
